Reimplementation with essential player
The player now is essential, and their death is managed through the
OnEnterBleedout event. A few changes are needed to make the animations
for bleedout the ones associated with death.

This reimplementation is necessary so that the time intervals for
different deaths (quick deaths, normal deaths, dynamics deaths...) can
stay the same (or be configured). This is also important for the "pless
key to reload" planned feature.